api user

+++ src/main/java/com/example/doan/service/impl/AuthServiceImpl.java
@@ -35,22 +35,25 @@ public class AuthServiceImpl implements AuthService {
     private final EmailService emailService;
     private final VerificationCodeRepository verificationCodeRepository;
     private final UserRepository userRepository;
-    private final customerUserServiceImpl customerUserService;
+    private final CustomerUserServiceImpl customerUserService;
 
     @Override
-    public void sentLoginOtp(String email,String password, USER_ROLE role) throws Exception {
+    public void sentLoginOtp(String email, String password, USER_ROLE role) throws Exception {
         String SIGNING_PREFIX = "signing_";
 
         if (email.startsWith(SIGNING_PREFIX)) {
             email = email.substring(SIGNING_PREFIX.length());
-
-            // Giả sử bạn chỉ còn User role là ADMIN và CUSTOMER
-            User user = userRepository.findByEmail(email);
-            if (user == null) {
-                throw new Exception("User không tồn tại với email đã cung cấp");
-            }
         }
-        // Xử lý OTP
+
+        User user = userRepository.findByEmail(email);
+        if (user == null) {
+            throw new Exception("User không tồn tại với email đã cung cấp");
+        }
+
+        if (!this.passwordEncoder.matches(password, user.getPassword())) {
+            throw new Exception("Mật khẩu không đúng");
+        }
+
         VerificationCode existing = verificationCodeRepository.findByEmail(email);
         if (existing != null) {
             verificationCodeRepository.delete(existing);
@@ -62,7 +65,7 @@ public class AuthServiceImpl implements AuthService {
         verificationCode.setEmail(email);
         verificationCodeRepository.save(verificationCode);
 
-        String subject = "Mã OTP đăng nhập/đăng ký";
+        String subject = "Mã OTP đăng nhập";
         String text = "Mã OTP của bạn là: " + otp;
 
         emailService.sendVerificationEmail(email, otp, subject, text);
@@ -71,20 +74,20 @@ public class AuthServiceImpl implements AuthService {
     @Override
     public String createUser(SignupRequest req) throws Exception {
 
-        VerificationCode verificationCode = verificationCodeRepository.findByEmail(req.getEmail());
-
-        if (verificationCode == null || !verificationCode.getOtp().equals(req.getOtp())) {
-            throw new Exception("wrong otp!");
-        }
+//        VerificationCode verificationCode = verificationCodeRepository.findByEmail(req.getEmail());
+//
+//        if (verificationCode == null || !verificationCode.getOtp().equals(req.getOtp())) {
+//            throw new Exception("wrong otp!");
+//        }
         User user = userRepository.findByEmail(req.getEmail());
 
         if (user == null) {
             User createUser = new User();
             createUser.setEmail(req.getEmail());
             createUser.setFullName(req.getFullName());
+            createUser.setPhone(req.getPhone());
             createUser.setRole(USER_ROLE.ROLE_CUSTOMER);
-            createUser.setPhone("0888291733");
-            createUser.setPassword(passwordEncoder.encode(req.getOtp()));
+            createUser.setPassword(passwordEncoder.encode(req.getPassword()));
 
             user= userRepository.save(createUser);
 
@@ -119,6 +122,46 @@ public class AuthServiceImpl implements AuthService {
         authResponse.setRole(USER_ROLE.valueOf(roleName));
 
         return authResponse;
+    }
+
+    @Override
+    public void forgotPassword(String email) throws Exception {
+        User user = userRepository.findByEmail(email);
+        if (user == null) throw new Exception("Không tìm thấy người dùng với email");
+
+
+        VerificationCode existing = verificationCodeRepository.findByEmail(email);
+        if (existing != null) {
+            verificationCodeRepository.delete(existing);
+        }
+
+        String otp = OtpUtils.generateOtp();
+        VerificationCode verificationCode = new VerificationCode();
+        verificationCode.setOtp(otp);
+        verificationCode.setEmail(email);
+        verificationCodeRepository.save(verificationCode);
+
+        String subject = "Khôi phục mật khẩu";
+        String text = "Mã OTP đặt lại mật khẩu của bạn là: " + otp;
+        emailService.sendVerificationEmail(email, otp, subject, text);
+
+    }
+
+    @Override
+    public void resetPasswordOtp(String email, String otp, String newPassword) throws Exception {
+        VerificationCode verificationCode = verificationCodeRepository.findByEmail(email);
+        if (verificationCode == null || !verificationCode.getOtp().equals(otp)) {
+            throw new Exception("OTP không đúng hoặc đã hết hạn");
+        }
+
+        User user = userRepository.findByEmail(email);
+        if (user == null) throw new Exception("Không tìm thấy người dùng");
+
+        user.setPassword(passwordEncoder.encode(newPassword));
+        userRepository.save(user);
+
+        verificationCodeRepository.delete(verificationCode); // Xoá mã OTP sau khi dùng
+
     }
 
     private Authentication authenticate(String username, String otp) throws Exception {
